completed authentication, solved type & lint errors

+++ src/actions/actions.ts
@@ -1,18 +1,123 @@
 "use server";
 
+import { auth, signIn, signOut } from "@/lib/auth";
 import prisma from "@/lib/db";
-import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
-import { petFormSchema, petIdSchema } from "@/lib/validations";
-import { Pet } from "@prisma/client";
+import { authSchema, petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
+import bcrypt from "bcryptjs"
+import { checkAuth, getPetById } from "@/lib/server-utils";
+import { Prisma } from "@prisma/client";
+import { AuthError } from "next-auth";
 
-export async function addPet(pet : unknown) {
+// --- user actions ---
+export async function logIn( prevState : unknown, formData: unknown) {
+
+    await sleep(1000);
+    // converting form data into normal js objects
+    // it is similar to -> email = formData.get("email")
+
+    // check if formData is formData type
+    if(!(formData instanceof FormData)){
+        return {
+            message: "Invalid form data"
+        }
+    }
+
+    // convert formData to an Object
+    // const formDataObject = Object.fromEntries(formData.entries())
+
+    // validate the object
+    
+    try {
+        await signIn("credentials", formData)
+        
+    } catch (error) {
+        if(error instanceof AuthError){
+            switch(error.type){
+                case "CredentialsSignin" : {
+                    return {
+                        message : "Invalid credentials"
+                    }
+                }
+                default : {
+                    return {
+                        message: "Error. Could not sign in"
+                    }
+                }
+            }
+        }
+        throw error; // nextjs redirects throws error, so we need to rethrow it
+    }
+}
+
+export async function logOut() {
+    await sleep(1000);
+
+    await signOut({ redirectTo: "/" })
+}
+
+export async function signUp(prevState : unknown, formData: unknown) {
+
+    await sleep(1000);
+    // check if formData is formData type
+    if(!(formData instanceof FormData)){
+        return {
+            message : "Invalid form data"
+        }
+    }
+
+    // convert formData to plain object
+    const formDataEntries = Object.fromEntries(formData.entries());
+
+    // validation
+    const validatedFormData = authSchema.safeParse(formDataEntries)
+    if(!validatedFormData.success){
+        return {
+            message: "Invalid form data"
+        }
+    }
+
+    const {email , password} = validatedFormData.data
+
+    const hashedPassword = await bcrypt.hash(password, 10)
+
+    try {
+        await prisma.user.create({
+            data: {
+                email,
+                hashedPassword,
+            }
+        })
+    } catch (error) {
+        if(error instanceof Prisma.PrismaClientKnownRequestError){
+            if(error.code === "P2002"){
+                return {
+                    message: "Email already exists"
+                }
+            }
+        }
+
+        return {
+            message: "Could not create user"
+        }
+    }
+
+    await signIn("credentials", formData)
+
+}
+
+
+
+// --- pet actions ---
+export async function addPet(pet: unknown) {
 
     await sleep(1000)
 
+    const session = await checkAuth()
+
     const validatedPet = petFormSchema.safeParse(pet)
-    if(!validatedPet.success){
+    if (!validatedPet.success) {
         return {
             message: "Invalid pet data",
         };
@@ -20,7 +125,14 @@ export async function addPet(pet : unknown) {
 
     try {
         await prisma.pet.create({
-            data: validatedPet.data
+            data: {
+                ...validatedPet.data,
+                user: {
+                    connect: {
+                        id: session.user.id
+                    }
+                }
+            }
         })
     } catch (error) {
         return {
@@ -31,18 +143,36 @@ export async function addPet(pet : unknown) {
 }
 
 
-export async function editPet(petId : unknown, newPetData : unknown) {
+export async function editPet(petId: unknown, newPetData: unknown) {
     await sleep(1000);
 
-		const validatedId = petIdSchema.safeParse(petId) 
+    // authentication check
+    const session = await checkAuth()
 
-		const validatedPet = petFormSchema.safeParse(newPetData)
-    if(!validatedPet.success || !validatedId.success){
+    // validation
+    const validatedId = petIdSchema.safeParse(petId)
+
+    const validatedPet = petFormSchema.safeParse(newPetData)
+    if (!validatedPet.success || !validatedId.success) {
         return {
             message: "Invalid pet data",
         };
     }
 
+    // authorization check
+    const pet = await getPetById(validatedId.data)
+    if(!pet){
+        return {
+            message : "Pet not found"
+        }
+    }
+    if(pet.userId !== session.user.id){
+        return {
+            message : "Not authorized"
+        }
+    }
+
+    // database mutation
     try {
         await prisma.pet.update({
             where: {
@@ -58,20 +188,39 @@ export async function editPet(petId : unknown, newPetData : unknown) {
     revalidatePath("/app", "layout")
 }
 
-export async function deletePet(petId : unknown) {
+export async function deletePet(petId: unknown) {
     await sleep(1000)
 
-		const validatedId = petIdSchema.safeParse(petId)
-		if(!validatedId.success){
-			return {
-				message : "Invalid pet data"
-			}
-		}
+    // authentication check
+    const session = await checkAuth()
 
-    try{
+    // validation
+    const validatedPetId = petIdSchema.safeParse(petId)
+    if (!validatedPetId.success) {
+        return {
+            message: "Invalid pet data"
+        }
+    }
+
+    // authorization check 
+    const pet = await getPetById(validatedPetId.data)
+
+    if(!pet){
+        return {
+            message : "Pet not found"
+        }
+    }
+    if(pet.userId !== session.user.id){
+        return {
+            message : "Not authorized"
+        }
+    }
+
+    // database mutation
+    try {
         await prisma.pet.delete({
             where: {
-                id : validatedId.data
+                id: validatedPetId.data
             }
         })
     } catch (error) {
